requirements.txt
Flask
Pillow
jsmin
cssmin
